@@ -1,11 +1,11 @@
 package WindGirls.MoneyMinder.domain;
 
 
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @Getter
